deleted generated RPC files awgtpman program compiled, runs and both tpman and AWG functionality works (tested with diag -l)
git-svn-id: https://redoubt.ligo-wa.caltech.edu/svn/advLigoRTS/trunk@1949 6dcd42c9-f523-4c6d-aada-af552506706e

+++ src/gds/rchannel.h
@@ -6,7 +6,9 @@
 #ifndef _RCHANNEL_H_RPCGEN
 #define _RCHANNEL_H_RPCGEN
 
+#include <rpc/rpc.h>
 
+#include <pthread.h>
 
 #ifdef __cplusplus
 extern "C" {
@@ -22,7 +24,7 @@ struct channelinfo_r {
 	short dcuId;
 	short chNum;
 	short dataType;
-	short dataRate;
+	int dataRate;
 	short chGroup;
 	short bps;
 	float gain;
